Moved `PropertyAccess` classes to model package
