+++ ui/de-lib/src/main/java/org/iplantc/de/admin/desktop/client/toolAdmin/view/ToolAdminDetailsView.ui.xml
@@ -1,7 +1,6 @@
 <!DOCTYPE ui:UiBinder SYSTEM "http://dl.google.com/gwt/DTD/xhtml.ent">
 <ui:UiBinder xmlns:ui="urn:ui:com.google.gwt.uibinder"
              xmlns:container="urn:import:com.sencha.gxt.widget.core.client.container"
-             xmlns:gxt="urn:import:com.sencha.gxt.widget.core.client"
              xmlns:g="urn:import:com.google.gwt.user.client.ui"
              xmlns:d="urn:import:org.iplantc.de.admin.desktop.client.toolAdmin.view.subviews"
              xmlns:form="urn:import:com.sencha.gxt.widget.core.client.form">
