+++ ui/de-lib/src/main/java/org/iplantc/de/admin/desktop/client/toolAdmin/view/subviews/ToolTestDataEditor.ui.xml
@@ -4,8 +4,7 @@
              xmlns:g="urn:import:com.google.gwt.user.client.ui"
              xmlns:d="urn:import:org.iplantc.de.admin.desktop.client.toolAdmin.view.subviews"
              xmlns:toolbar="urn:import:com.sencha.gxt.widget.core.client.toolbar"
-             xmlns:button="urn:import:com.sencha.gxt.widget.core.client.button"
-        >
+             xmlns:button="urn:import:com.sencha.gxt.widget.core.client.button">
 
     <ui:with field="appearance"
              type="org.iplantc.de.admin.desktop.client.toolAdmin.ToolAdminView.ToolAdminViewAppearance"/>
